finetune/requirements.txt
torch
transformers>=4.44.0
datasets>=2.19.0
peft>=0.11.1
accelerate>=0.33.0
bitsandbytes>=0.43.0
sentencepiece>=0.2.0
tokenizers>=0.19.0
pyyaml>=6.0.1
safetensors>=0.4.3
evaluate>=0.4.2
tqdm>=4.66.0
protobuf>=4.25.0
einops>=0.8.0
